+++ terraform/modules/s3/outputs.tf
@@ -1,0 +1,11 @@
+output "bucket_name" {
+  description = "S3 bucket name."
+
+  value = aws_s3_bucket.assets.bucket
+}
+
+output "bucket_arn" {
+  description = "S3 bucket ARN."
+
+  value = aws_s3_bucket.assets.arn
+}